feat: entidades jpa, repositorios y auditing con lombok
- Lombok 1.18.46 con annotation processor en maven-compiler-plugin
- Auditable @MappedSuperclass con @CreatedDate, @LastModifiedDate,
  @CreatedBy, @LastModifiedBy
- AuditorAwareImpl: devuelve "system" como fallback (Fase 5 lo
  conecta al SecurityContext)
- JpaConfig habilita @EnableJpaAuditing
- 9 entidades JPA con jakarta.persistence + jakarta.validation:
  Perfil, Usuario, Parametro, Contacto, Menu, Submenu,
  SubmenuPerfilId (@Embeddable), SubmenuPerfil (@EmbeddedId +
  @MapsId), RefreshToken
- Relaciones LAZY, equals/hashCode por id (callSuper=false)
- 8 repositorios con metodos derivados y JPQL findMenuPorPerfil
- application-test.yml apunta a baqueano_test
- scripts/db_setup.sql ahora tambien crea baqueano_test
- V1__schema_inicial.sql: intentos_fallidos pasa de TINYINT a INT
  (Java Integer no mapea a TINYINT por defecto en Hibernate 6.6)
- 21 tests con @DataJpaTest + @AutoConfigureTestDatabase(NONE) +
  @ActiveProfiles(test), 100% verde

+++ baqueano-backend/src/main/resources/db/migration/V1__schema_inicial.sql
@@ -30,7 +30,7 @@ CREATE TABLE usuarios (
     perfil_id       BIGINT       NOT NULL,
     activo          BOOLEAN      NOT NULL DEFAULT TRUE,
     ultimo_login    DATETIME     NULL,
-    intentos_fallidos TINYINT    NOT NULL DEFAULT 0,
+    intentos_fallidos INT        NOT NULL DEFAULT 0,
     fecha_creacion       DATETIME NOT NULL DEFAULT CURRENT_TIMESTAMP,
     fecha_modificacion   DATETIME NOT NULL DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP,
     usuario_creacion     VARCHAR(50),
